requirements.txt
keystone-engine~=0.9.2
pyfastboot~=1.3.11
